HSEARCH-3440 Move some test methods from LuceneLocalDirectoryIT to an abstract class
